Create nginix-installation.sh
This script will first check if NGINX is already installed. If it is, the script will exit without doing anything. If NGINX is not installed, the script will install the NGINX package, start the NGINX service, enable the NGINX service to start at boot, and check the status of the NGINX service. Finally, the script will print a message to the user indicating that NGINX has been successfully installed.

To run this script, you can save it as a .sh file and then run it from the command line. For example, if you saved the script as install_nginx.sh, you could run it by typing the following command:

touch install_nginx.sh

chmod+x install_nginx.sh

./install_nginx.sh