check-web-app.sh
The first line of the code, #!/bin/bash, tells the shell that the script is written in Bash.

The next two lines set the URL of the web application to check and the email address to send alerts to.

The curl command is used to check the status of the web application. The -s option tells curl to be silent, the -o /dev/null option tells curl to write the output to a null device, and the -w '%{http_code}\n' option tells curl to print the HTTP status code to the standard output.

The if statement checks if the response from curl is not 200. If the response is not 200, the script will send an alert to the email address specified.

The echo commands are used to print messages to the standard output. The mail command is used to send an email.
Here is a step-by-step explanation of what the code does:

The first line of the code tells the shell that the script is written in Bash.
The second line sets the URL of the web application to check.
The third line sets the email address to send alerts to.
The fourth line uses the curl command to check the status of the web application. The -s option tells curl to be silent, the -o /dev/null option tells curl to write the output to a null device, and the -w '%{http_code}\n' option tells curl to print the HTTP status code to the standard output.
The fifth line uses the if statement to check if the response from curl is not 200. If the response is not 200, the script will send an alert to the email address specified.
The sixth line uses the echo command to print a message to the standard output.
The seventh line uses the mail command to send an email.

In the context of HTTP, 200 is the status code for a successful request. It means that the request was received and understood, and the server is returning the requested resource.

The HTTP status code 200 is also known as "OK". It is the most common status code, and it is used for a variety of requests, including GET, POST, PUT, and DELETE.

If you see a 200 status code in your browser's status bar, it means that the request was successful and the resource was returned.

Here are some other common HTTP status codes:

400 Bad Request: The request was malformed or could not be understood.
401 Unauthorized: The request requires authentication.
403 Forbidden: The request is forbidden.
404 Not Found: The requested resource was not found.
500 Internal Server Error: The server encountered an error while processing the request.

To run this script, you will need to save it as a file with the .sh extension and then run it using the bash command. For example, to run the script from the current directory, you would run the following command:

bash check-web-app.sh
This will check the status of the web application and send an alert if it is down.